CAPI-235(chore): move configmanager to its own directory
